+++ packages/cloudflare-tunnel-client/src/index.ts
@@ -1,0 +1,3 @@
+export { CloudflareTunnelClient } from "./CloudflareTunnelClient.js";
+export { SubscriptionValidator } from "./SubscriptionValidator.js";
+export * from "./types.js";